+++ src/cvc/helperFunctions.py
@@ -1,13 +1,35 @@
+import time
 import requests
 from bs4 import BeautifulSoup, Tag
+
+from selenium import webdriver
+from selenium.webdriver.common.by import By
+from selenium.webdriver.support.ui import WebDriverWait
+from selenium.webdriver.support import expected_conditions as EC
+from selenium.webdriver.common.keys import Keys
+from selenium.webdriver.common.action_chains import ActionChains
+from selenium.webdriver.chrome.service import Service
+# from webdriver_manager.chrome import ChromeDriverManager
+
+import re
+import json
 
 from .course import course
 from .courseSection import courseSection
 
+from .sharedVariables import _getGreatestMatching
+
+headers = {
+    "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/112.0.0.0 Safari/537.36",
+    "Accept-Language": "en-US,en;q=0.9",
+    "Referer": "https://google.com", 
+    "Accept": "text/html,application/xhtml+xml",
+}
+
 def _getPageContentFromID(id:int):
     url = f'https://search.cvc.edu/courses/{id}'
     # Send GET request to the URL
-    response = requests.get(url)
+    response = requests.get(url, headers=headers)
 
     if response.status_code != 200:
         raise ValueError("Inputted ID is invalid.", id)
@@ -20,7 +42,7 @@ def _getPageContentFromID(id:int):
         return ""
     return soup
 
-def _parseCourse(soup):
+def _parseCourse(soup,id:int):
     table = soup.find('div', attrs = {'class':'font-semibold text-sm md:text-base'})
     try:
         college_name_text = table.get_text(strip=True).strip()
@@ -46,23 +68,56 @@ def _parseCourse(soup):
     except (TypeError, AttributeError, ValueError) as e:
         C_ID = ""
     
+    table = soup.find('h2', string='Location',attrs={'class':'font-semibold text-sm mb-1 uppercase text-c_content_text_highlight'})
+    table = table.find_next_siblings()[0]
+    locationText = table.get_text(strip=True)
+
+    table = soup.find('h2', string='Units',attrs={'class':'font-semibold text-sm mb-1 uppercase text-c_content_text_highlight'})
+    table = table.find_next_siblings()[0]
+    unitText = table.get_text(strip=True)
+    if unitText:
+        match = re.search(r'(\d+(\.\d+)?)\s+(semester|quarter)', unitText, re.IGNORECASE)
+        if match:
+            units = match.group(1)                  # i.e. "4.0"
+            termType = match.group(3).capitalize()  # i.e. "semester"
+        else:
+            units = unitText
+            termType = unitText
+    else:
+        units = ""
+        termType = ""
+
     table = soup.find('div', attrs = {'class':'course-description pb-2'}) 
-    try:
-        course_description = table.get_text(strip=True)
-        index = course_description.index('Prerequisite:')
-        prereqs = course_description[index+len("Prerequisite:")+1:]
-        course_description = course_description[:index]
-        index = prereqs.index('all with')
-        prereqs = prereqs[:index].replace(",", "").replace(".", "").replace("and","").replace("or","").replace("-", " ")
+    coureDescriptionText = table.get_text(strip=True)
+    match = re.search(r'(Prerequisite:.*?(\.|\n))', coureDescriptionText)
+    prereqs = []
+    if match:
+        prerequisite = match.group(1)  # Store the matched sentence
+        prereqs = prerequisite.replace(",", "").replace(".", "").replace("and","").replace("or","").replace("-", " ").replace("Prerequisites:","").replace("Prerequisite:","")
         prereqs = prereqs.split("  ")
-        for i in range(0,len(prereqs)):
-            prereqs[i] = prereqs[i].strip()
-        #print(course_description)
-        #print(prereqs)
-    except (TypeError, AttributeError, ValueError) as e:
-        course_description = ""
+        coureDescriptionText = table.get_text(strip=True).replace(prerequisite, '').strip()  # Remove it from the text
+    else:
         prereqs = []
-    return course(college_name_text,class_name_text,class_symbol,C_ID,course_description,prereqs,_parseSections(soup))
+    # try:
+    #     course_description = table.get_text(separator=' ',strip=True)
+    # except (TypeError, AttributeError, ValueError) as e:
+    #     print(e)
+    #     course_description = ""
+    # try:
+    #     index = course_description.index('Prerequisite:')
+    #     if index > -1:
+    #         course_description = course_description[:index]
+    #     prereqs = course_description[index+len("Prerequisite:")+1:]
+    #     index = prereqs.index('all with')
+    #     prereqs = prereqs[:index].replace(",", "").replace(".", "").replace("and","").replace("or","").replace("-", " ")
+    #     prereqs = prereqs.split("  ")
+    #     for i in range(0,len(prereqs)):
+    #         prereqs[i] = prereqs[i].strip()
+    #     #print(course_description)
+    #     #print(prereqs)
+    # except (TypeError, AttributeError, ValueError) as e:
+    #     prereqs = []
+    return course(id, college_name_text,class_name_text,class_symbol,C_ID,coureDescriptionText,locationText, units, termType,prereqs,_parseSections(soup))
 
 def _parseSections(soup):
     table = soup.find('h3', attrs={'class':'font-semibold mb-4 border-b-2 border-c_content_text_highlight pb-2'})
@@ -96,10 +151,20 @@ def _parseSections(soup):
                 formatText = format.get_text(strip=True)
                 format.extract()
 
-                time = sem.find('span', attrs={'class': 'text-xs font-medium'})
-                timeText = time.get_text(strip=True)
-                time.extract()
-
+                time = sem.find('span', text='Time:',attrs={'class': 'font-semibold text-c_content_text_highlight text-xs mr-1'})
+                neighbors = list(time.find_next_siblings())
+                times = []
+                
+                for n in neighbors:
+                    if n.name == 'ul':
+                        li_items = n.find_all('li', recursive=False)
+                        for li in li_items:
+                            times.append(li.get_text(strip=True))
+                        continue
+                    times.append(n.get_text(strip=True))
+                if time:
+                    time.extract()
+                
                 prof = sem.find('span', attrs={'class': 'text-xs font-medium section-details-professor'})
                 profText = prof.get_text(strip=True)
                 prof.extract()
@@ -111,7 +176,7 @@ def _parseSections(soup):
                 seatCountText = seatCountText[:index]
 
                 sectionNotes = sem.find('div', attrs={'class': 'text-xs section-details-notes'})
-                sectionNotesText = sectionNotes.get_text(strip=True)
+                sectionNotesText = sectionNotes.get_text(separator=' ',strip=True)
                 sectionNotes.extract()
 
                 tuition = sem.find('span', attrs={'class': 'font-semibold text-lg text-c_link'})
@@ -125,12 +190,145 @@ def _parseSections(soup):
                 if zeroTextbookCostItem:
                     zeroTextbookCost = True
                     zeroTextbookCostItem.extract()
+                
+                registration = ""
+                startingDate = sem.find('span', text=' Already Started ',attrs={'class': 'text-sm font-semibold'})
+                if startingDate:
+                    registration = "Already Open"
+                else:    
+                    startingDate = sem.find('span',attrs={'class': 'text-sm font-semibold'})
+                    if startingDate != None:
+                        startingDate = startingDate.find('span', attrs={'class':'text-c_content_text_highlight'})
+                        if startingDate != None:
+                            registration = startingDate.get_text(strip=True).replace(" Registration Opens ","")
 
-                foundSection = courseSection(currentSem,sectionIdText,durationText,formatText,zeroTextbookCost,timeText,profText,seatCountText,tuitionText,sectionNotesText)
+
+                foundSection = courseSection(currentSem,durationText,sectionIdText,formatText,zeroTextbookCost,times,profText,seatCountText,tuitionText,registration,sectionNotesText)
                 #foundSection.printAll()
                 sections.append(foundSection)
     return sections
 
 def _getCourseContentByID(id:int):
     soup = _getPageContentFromID(id)
-    return _parseCourse(soup)
+    return _parseCourse(soup,id)
+
+def _getCourseIDsByLandingPage(collegeName:str, C_ID:str, courseSymbol:str, courseName:str):
+    maxScore = 0
+    if C_ID != "":
+        maxScore+=1
+        C_ID = C_ID.lower().strip()
+    if courseSymbol != "":
+        maxScore += 1
+        courseSymbol = courseSymbol.lower().strip()
+    if courseName != "":
+        maxScore += 1
+        courseName = courseName.lower().strip()
+    driver = _getDriver()
+    driver.get('https://search.cvc.edu/embedded_search') 
+
+    WebDriverWait(driver, 20).until(
+        lambda driver: driver.execute_script('return document.readyState') == 'complete'
+    )
+    WebDriverWait(driver, 20).until(
+        lambda driver: driver.execute_script('return window.jQuery && jQuery.active == 0')
+    )
+    # with open("page_source.html", "w", encoding="utf-8") as file:
+    #     file.write(driver.page_source)    
+    radio_btn = WebDriverWait(driver, 20).until(
+        EC.presence_of_element_located((By.NAME, "filter[search_all_universities]"))
+    )
+    input_box = WebDriverWait(driver, 20).until(
+        EC.presence_of_element_located((By.ID, "filter_university_id-selectized"))
+    )
+    input_box.send_keys(collegeName, Keys.RETURN)
+
+    HomeCollegeText = driver.find_element(By.XPATH, "//span[text()='Home College Course Name']").find_element(By.XPATH, '..').find_element(By.XPATH, ".//input[@id='filter_search_type_course']")
+    actions = ActionChains(driver)
+    actions.move_to_element(HomeCollegeText).click().perform()
+    HomeCollegeText.click()
+
+    input_box = WebDriverWait(driver, 20).until(
+        EC.presence_of_element_located((By.ID, "filter_query-selectized"))
+    )
+    # print((courseSymbol + " " + "(C-ID: "+C_ID + ") " + courseName).strip())
+    input_box.send_keys((courseSymbol + " " + C_ID + " " + courseName).strip())
+    time.sleep(1)
+    input_box.send_keys(Keys.RETURN)
+    html = driver.page_source
+    soup = BeautifulSoup(html, 'html.parser')
+    courses = soup.find_all('div', attrs={'class':'option'})
+    courseIDs = []
+    for course in courses:
+        score = 0
+        if("course-" not in course.get('id')):
+            continue
+        courseID = course.get('id').replace("course-","")
+        result = course.get_text().lower()
+        result = result.replace(" Show substitutes for ","")
+        result = result.strip()
+        if C_ID != "":
+            match = re.search(r"\((.*?)\)", result)
+            if match:
+                foundC_ID = match.group(1)
+                foundC_ID = foundC_ID.replace("(C-ID: ","").replace(")","").lower().strip()
+                if C_ID in foundC_ID:
+                    score += 1
+        if maxScore == score:
+            courseIDs.append(courseID)
+            continue
+        result = re.sub(r"\(.*?\)", "", result)
+
+        if courseSymbol != "":
+            if courseSymbol in result:
+                result = result.replace(courseSymbol, "")
+                score +=1
+        if maxScore == score:
+            courseIDs.append(courseID)
+            continue
+        
+        if courseName != "":
+            if courseName in result:
+                score +=1
+        if maxScore == score:
+            courseIDs.append(courseID)
+            continue
+    if len(courseIDs) > 0:
+        return courseIDs
+    else:
+        print("No course could be found with these parameters.")
+        return []
+    driver.quit()
+
+def _getDriver():
+    """
+    Creates a webdriver to be used with Selenium to scrape a page. So far used in _getCourseIDsByLandingPage.
+
+    Returns:
+    webdriver: A Chrome browser which will handle automated interaction.
+    """
+    options = webdriver.ChromeOptions()
+    options.add_argument("--log-level=3")  # Suppresses INFO and WARNING
+    options.add_argument('--no-sandbox')
+    options.add_argument('--disable-dev-shm-usage') 
+    options.add_argument('--headless')
+    options.add_experimental_option('excludeSwitches', ['enable-logging'])
+    options.set_capability("browserVersion", "117")
+    driver = webdriver.Chrome(options=options) #service=Service(ChromeDriverManager().install())
+    return driver
+
+def _getCourseIDsBySearch(collegeName:str, C_ID:str, courseSymbol:str, courseName:str):
+    #https://search.cvc.edu/api/search.json?query=cd&university_id=48&only_visible=true
+    #https://search.cvc.edu/api/search.json?query=cs&university_id=48&only_visible=true
+    #https://search.cvc.edu/api/search.json?query=fun%20stuff&university_id=48&only_visible=true
+    url = f'https://search.cvc.edu/api/search.json?query={(courseSymbol + " " + C_ID + " " + courseName).strip()}&university_id={_getGreatestMatching(collegeName)}&only_visible=false'
+    # Send GET request to the URL
+    response = requests.get(url, headers=headers)
+    if response.status_code != 200:
+        raise ValueError("No course could be found. Status code:", response.status_code)
+        return []
+    data = response.json()
+    data = data['results']
+    courseIDs = []
+    for course in data:
+        courseIDs.append(course['id'].replace("course-",""))
+    return courseIDs