+++ CHANGELOG.md
@@ -6,3 +6,9 @@ Change Log
 - First Release
 - Added the course and courseSection classes
 - Added the getCourseContentByID method
+
+0.0.2 (April 16th, 2025)
+-----------------------
+- Added new parameters to the course and courseSection classes
+- Added the getCourseContentByScraping, getCourseIDsByScraping, and getCourseIDsBySearch methods.
+- Updated documentation now posted on https://cvc-api.github.io/